Yilin Project 3 Code Review DB Connections Design
To make my opinion clear, this is not a true-or-false problem. Using a new connection for every request simplifies the logic and it's easy to manage the connection (just close it when the request is done in the back-end). But you cannot ignore the overhead of establishing a new connection in every request. While, when sharing the same connection in multiple requests, you do need to have a systematic way to manage the long-lived connections.