Yilin Project 3 Code Review DB Connection Design
To make my opinion clear, it's absolutely okay to share the connection. It's not a true-or-false problem. Sharing the connection saves the overhead of establishing a new connection in every request. However, it requires some further concerns like how to manage them. On the other hand, using a new connection in each request simplifies the logic but you cannot ignore the extra cost.